Refactor to use components: post content, sidebar, and comments components;
